+++ src/soc/amd/common/block/root_complex/non_pci_resources.c
@@ -7,7 +7,7 @@
 
 #define IOHC_MMIO_EN	BIT(0)
 
-void read_non_pci_resources(struct device *domain, unsigned int *idx)
+void read_non_pci_resources(struct device *domain, unsigned long *idx)
 {
 	const uint32_t iohc_misc_base = get_iohc_misc_smn_base(domain);
 	const struct non_pci_mmio_reg *regs;
